Fix localization file loading, and introduce the load_textdomain() method
As pointed out by aarcos in the forums
(http://wordpress.org/support/topic/translation-files-not-loading-and-suggested-patch-code#post-4754557),
localization files are not loading correctly, due to an invalid path argument.

Fix loading by correcting the path, and add more locations to check.
This allows users to add custom translation files that won't be stomped
out when the plugin is updated. Introduce the load_textdomain() method
to handle the more verbose setup.

git-svn-id: https://plugins.svn.wordpress.org/edit-author-slug/trunk@824522 b8457f37-d9ea-0310-8a92-e5e31aec5664